pc配置中心-参数库配置

+++ core/src/main/java/com/smart/home/modules/product/service/ProductParamSettingService.java
@@ -1,9 +1,13 @@
 package com.smart.home.modules.product.service;
 
 import com.github.pagehelper.PageHelper;
+import com.smart.home.common.exception.DuplicateDataException;
+import com.smart.home.common.exception.ServiceException;
 import com.smart.home.modules.product.dao.ProductParamSettingMapper;
+import com.smart.home.modules.product.dao.ProductParamValueMapper;
 import com.smart.home.modules.product.entity.ProductParamSetting;
 import com.smart.home.modules.product.entity.ProductParamSettingExample;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,13 +23,28 @@ public class ProductParamSettingService {
 
     @Resource
     ProductParamSettingMapper productParamSettingMapper;
+    @Resource
+    ProductParamValueMapper productParamValueMapper;
 
     public int create(ProductParamSetting productParamSetting) {
+        ProductParamSettingExample example = new ProductParamSettingExample();
+        example.createCriteria().andParamNameEnEqualTo(productParamSetting.getParamName());
+        if (productParamSettingMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该参数已经存在");
+        }
+        productParamSetting.setRevision(0);
         productParamSetting.setCreatedTime(new Date());
         return productParamSettingMapper.insertSelective(productParamSetting);
     }
 
     public int update(ProductParamSetting productParamSetting) {
+        ProductParamSettingExample example = new ProductParamSettingExample();
+        example.createCriteria()
+                .andParamNameEnEqualTo(productParamSetting.getParamName())
+                .andIdNotEqualTo(productParamSetting.getId());
+        if (productParamSettingMapper.countByExample(example) > 0) {
+            throw new DuplicateDataException("该参数已经存在");
+        }
         productParamSetting.setUpdatedTime(new Date());
         return productParamSettingMapper.updateByPrimaryKeySelective(productParamSetting);
     }
@@ -37,6 +56,10 @@ public class ProductParamSettingService {
     @Transactional(rollbackFor = RuntimeException.class)
     public void delete(List<Long> idList) {
         for (Long id : idList) {
+            // 判断有没有关联产品了，关联产品后不能删除
+            if (productParamValueMapper.countByParamId(id.intValue()) > 0) {
+                throw new ServiceException("已经关联了产品，不能删除");
+            }
             productParamSettingMapper.deleteByPrimaryKey(id.intValue());
         }
     }
@@ -45,7 +68,9 @@ public class ProductParamSettingService {
         PageHelper.startPage(pageNum, pageSize);
         ProductParamSettingExample example = new ProductParamSettingExample();
         ProductParamSettingExample.Criteria criteria = example.createCriteria();
-        // TODO 按需根据字段查询
+        if (StringUtils.isNotBlank(productParamSetting.getParamName())) {
+            criteria.andParamNameEqualTo(productParamSetting.getParamName());
+        }
         return productParamSettingMapper.selectByExample(example);
     }
 
@@ -54,4 +79,8 @@ public class ProductParamSettingService {
         return productParamSetting;
     }
 
+    public List<ProductParamSetting> queryAllValid() {
+        ProductParamSettingExample example = new ProductParamSettingExample();
+        return productParamSettingMapper.selectByExample(example);
+    }
 }
